feat(stm32): blackbox logging to the sd card
ULog records cross from the control tick to the bench tier through a 32 KB
ring, then land as raw sectors inside a preallocated contiguous file, so FatFs
never runs while armed. USB mass storage and a FcLink read path serve the card
afterwards. Boot panics on an unusable card; a failure past init leaves the
service inert and counting.

+++ stm32/Services/crsf_link_service.cpp
@@ -262,7 +262,7 @@ CrsfLinkService::PrepareTelemetryTopic(TelemetryTopic topic,
     case TelemetryTopic::kGps: {
       const GpsData &gps = blackboard_->GetGps();
       if (gps.timestamp_us == 0 ||
-          (uint32_t)(now_us - gps.timestamp_us) > cfg_.gps_fresh_timeout_us) {
+          (now_us - gps.timestamp_us) > cfg_.gps_fresh_timeout_us) {
         return std::nullopt;
       }
       frame.type = kCrsfFrameTypeGps;
